fix: restore strict path-specific CCSaaS org and cluster gates
The OC-aligned shared-factory approach applied both validators to both the login id_token
and the bearer token with lenient-on-absence semantics. Against the 8.9 baseline that is a
regression: a public-API bearer token missing the clusterId claim was accepted (8.9
rejected it), and the login org gate was weakened from strict to lenient.

Reproduce 8.9's strict, path-specific split instead. The organization + role gate is strict
and wired only into idTokenDecoderFactory (login id_token, which carries the orgs claim). The
cluster-id gate is strict (CustomClaimValidator) and stays on the bearer TokenValidatorFactory
(which M2M tokens hit, carrying clusterId but no org membership). No shared lenient factory.
