+++ optimize/backend/src/test/java/io/camunda/optimize/rest/security/csl/OptimizeCloudSecurityConfigurationTest.java
@@ -13,11 +13,9 @@ import static org.mockito.Mockito.when;
 import io.camunda.optimize.service.util.configuration.ConfigurationService;
 import io.camunda.optimize.service.util.configuration.security.AuthConfiguration;
 import io.camunda.optimize.service.util.configuration.security.CloudAuthConfiguration;
-import io.camunda.security.api.model.config.oidc.OidcConfiguration;
 import io.camunda.security.core.port.in.OidcProviderConfigurationPort;
 import io.camunda.security.spring.oidc.TokenValidatorFactory;
 import java.time.Instant;
-import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -40,11 +38,14 @@ class OptimizeCloudSecurityConfigurationTest {
   private final OptimizeCloudSecurityConfiguration config =
       new OptimizeCloudSecurityConfiguration();
 
-  @Test
-  void shouldComposeOrgAndClusterValidatorsIntoTheFactory() {
+  private void withCloudConfig() {
     when(configurationService.getAuthConfiguration()).thenReturn(authConfiguration);
     when(authConfiguration.getCloudAuthConfiguration()).thenReturn(cloudAuthConfiguration);
-    when(cloudAuthConfiguration.getOrganizationId()).thenReturn("org-1");
+  }
+
+  @Test
+  void shouldGateBearerTokensOnClusterIdOnly() {
+    withCloudConfig();
     when(cloudAuthConfiguration.getClusterId()).thenReturn("cluster-1");
     when(oidcProviderConfigurationPort.getOidcAuthenticationConfigurations()).thenReturn(Map.of());
 
@@ -52,33 +53,20 @@ class OptimizeCloudSecurityConfigurationTest {
         config.tokenValidatorFactory(oidcProviderConfigurationPort, configurationService);
     final OAuth2TokenValidator<Jwt> validator = factory.createTokenValidator(clientRegistration());
 
-    assertThat(validator.validate(jwt("org-1", "analyst", "cluster-1")).hasErrors()).isFalse();
-    // Wrong cluster is rejected by the cluster validator.
-    assertThat(validator.validate(jwt("org-1", "analyst", "other")).hasErrors()).isTrue();
-    // Disallowed org role is rejected by the organization validator.
-    assertThat(validator.validate(jwt("org-1", "viewer", "cluster-1")).hasErrors()).isTrue();
+    assertThat(validator.validate(jwtWithClusterId("cluster-1")).hasErrors()).isFalse();
+    assertThat(validator.validate(jwtWithClusterId("other-cluster")).hasErrors()).isTrue();
+    // Strict: a bearer token without the cluster-id claim is rejected (8.9 baseline).
+    assertThat(validator.validate(jwtWithoutClusterId()).hasErrors()).isTrue();
   }
 
   @Test
-  void shouldProvideIdTokenDecoderFactory() {
-    final TokenValidatorFactory factory =
-        new TokenValidatorFactory(Map.of(), OidcConfiguration.DEFAULT_CLOCK_SKEW, List.of());
+  void shouldProvideIdTokenDecoderFactoryForTheLoginGate() {
+    withCloudConfig();
+    when(cloudAuthConfiguration.getOrganizationId()).thenReturn("org-1");
+    when(oidcProviderConfigurationPort.getOidcAuthenticationConfigurations()).thenReturn(Map.of());
 
-    assertThat(config.idTokenDecoderFactory(factory)).isInstanceOf(OidcIdTokenDecoderFactory.class);
-  }
-
-  private static Jwt jwt(final String orgId, final String role, final String clusterId) {
-    final Instant now = Instant.now();
-    return Jwt.withTokenValue("token")
-        .header("alg", "none")
-        .issuedAt(now)
-        .expiresAt(now.plusSeconds(300))
-        .subject("user")
-        .claim(
-            OptimizeCloudOrganizationValidator.ORGANIZATIONS_CLAIM,
-            List.of(Map.of("id", orgId, "roles", List.of(role))))
-        .claim(OptimizeCloudClusterValidator.CLUSTER_ID_CLAIM, clusterId)
-        .build();
+    assertThat(config.idTokenDecoderFactory(oidcProviderConfigurationPort, configurationService))
+        .isInstanceOf(OidcIdTokenDecoderFactory.class);
   }
 
   private static ClientRegistration clientRegistration() {
@@ -89,5 +77,22 @@ class OptimizeCloudSecurityConfigurationTest {
         .authorizationUri("http://idp/authorize")
         .tokenUri("http://idp/token")
         .build();
+  }
+
+  private static Jwt jwtWithClusterId(final String clusterId) {
+    return baseJwt().claim(OptimizeCloudSecurityConfiguration.CLUSTER_ID_CLAIM, clusterId).build();
+  }
+
+  private static Jwt jwtWithoutClusterId() {
+    return baseJwt().build();
+  }
+
+  private static Jwt.Builder baseJwt() {
+    final Instant now = Instant.now();
+    return Jwt.withTokenValue("token")
+        .header("alg", "none")
+        .issuedAt(now)
+        .expiresAt(now.plusSeconds(300))
+        .subject("user");
   }
 }
